requirements.txt
django
djangorestframework
django-cors-headers 
joblib
numpy 
pandas 
seaborn 
matplotlib